feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task (#48799)
* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

* feat: [CDS-54471]: NG: Implement rollback logic from new Perpetual Task to old Perpetual Task

+++ 126-instance/src/main/java/io/harness/service/instancesync/InstanceSyncServiceImpl.java
@@ -8,6 +8,7 @@
 package io.harness.service.instancesync;
 
 import static io.harness.connector.ConnectorModule.DEFAULT_CONNECTOR_SERVICE;
+import static io.harness.data.structure.EmptyPredicate.isEmpty;
 import static io.harness.exception.WingsException.USER;
 
 import static java.util.Objects.isNull;
@@ -294,7 +295,7 @@ public class InstanceSyncServiceImpl implements InstanceSyncService {
           instanceSyncPerpetualTaskInfoDTO.setPerpetualTaskIdV2(
               instanceSyncPerpetualTaskMappingDTO.getPerpetualTaskId());
           instanceSyncPerpetualTaskInfoDTO =
-              instanceSyncPerpetualTaskInfoService.updatePerpetualTaskIdV2(instanceSyncPerpetualTaskInfoDTO);
+              instanceSyncPerpetualTaskInfoService.updatePerpetualTaskIdV1OrV2(instanceSyncPerpetualTaskInfoDTO);
         }
         return instanceSyncPerpetualTaskInfoDTO;
       }
@@ -453,14 +454,37 @@ public class InstanceSyncServiceImpl implements InstanceSyncService {
               "Infrastructure mapping not found for {}", instanceSyncPerpetualTaskInfoDTO.getInfrastructureMappingId());
           // delete instance sync perpetual task info record
           instanceSyncHelper.cleanUpOnlyInstanceSyncPerpetualTaskInfo(instanceSyncPerpetualTaskInfoDTO);
-          return;
+          continue;
+        }
+
+        if (isEmpty(instanceSyncPerpetualTaskInfoDTO.getDeploymentInfoDetailsDTOList())) {
+          // There is no deployment info left to process for instance sync
+          instanceSyncPerpetualTaskInfoService.deleteById(
+              instanceSyncPerpetualTaskInfoDTO.getAccountIdentifier(), instanceSyncPerpetualTaskInfoDTO.getId());
+          log.info("Deleted instance sync perpetual task info : {} as there is no deployment info to do instance sync",
+              instanceSyncPerpetualTaskInfoDTO.getId());
+          continue;
+        }
+
+        AbstractInstanceSyncHandler abstractInstanceSyncHandler =
+            instanceSyncHandlerFactoryService.getInstanceSyncHandler(
+                instanceSyncPerpetualTaskInfoDTO.getDeploymentInfoDetailsDTOList()
+                    .get(0)
+                    .getDeploymentInfoDTO()
+                    .getType(),
+                infrastructureMappingDTO.get().getInfrastructureKind());
+
+        if (!abstractInstanceSyncHandler.isInstanceSyncV2Enabled()) {
+          migrateToInstanceSyncV1(
+              instanceSyncPerpetualTaskInfoDTO, infrastructureMappingDTO.get(), abstractInstanceSyncHandler);
+          continue;
         }
 
         if (!doSvcAndEnvExist(infrastructureMappingDTO.get())) {
           // as either or both of svc and env don't exist, we delete the instances
           instanceSyncPerpetualTaskMappingService.delete(accountIdentifier, perpetualTaskId);
           deleteInstances(infrastructureMappingDTO.get());
-          return;
+          continue;
         }
 
         InstanceSyncData instanceSyncData = instancesPerTask.get(taskInfoId);
@@ -491,12 +515,30 @@ public class InstanceSyncServiceImpl implements InstanceSyncService {
     }
   }
 
+  private void migrateToInstanceSyncV1(InstanceSyncPerpetualTaskInfoDTO instanceSyncPerpetualTaskInfoDTO,
+      InfrastructureMappingDTO infrastructureMappingDTO, AbstractInstanceSyncHandler abstractInstanceSyncHandler) {
+    if (instanceSyncPerpetualTaskInfoDTO.getPerpetualTaskId() == null) {
+      String perpetualTaskIdV1 =
+          instanceSyncPerpetualTaskService.createPerpetualTask(infrastructureMappingDTO, abstractInstanceSyncHandler,
+              instanceSyncPerpetualTaskInfoDTO.getDeploymentInfoDetailsDTOList()
+                  .stream()
+                  .map(DeploymentInfoDetailsDTO::getDeploymentInfoDTO)
+                  .collect(Collectors.toList()),
+              abstractInstanceSyncHandler.getInfrastructureOutcome(infrastructureMappingDTO.getInfrastructureKind(),
+                  instanceSyncPerpetualTaskInfoDTO.getDeploymentInfoDetailsDTOList().get(0).getDeploymentInfoDTO(),
+                  instanceSyncPerpetualTaskInfoDTO.getConnectorIdentifier()));
+      instanceSyncPerpetualTaskInfoDTO.setPerpetualTaskIdV2(null);
+      instanceSyncPerpetualTaskInfoDTO.setPerpetualTaskId(perpetualTaskIdV1);
+      instanceSyncPerpetualTaskInfoService.updatePerpetualTaskIdV1OrV2(instanceSyncPerpetualTaskInfoDTO);
+    }
+  }
+
   private void cleanupPerpetualTaskV1(InstanceSyncPerpetualTaskInfoDTO instanceSyncPerpetualTaskInfoDTO) {
     if (instanceSyncPerpetualTaskInfoDTO.getPerpetualTaskId() != null) {
       instanceSyncPerpetualTaskService.deletePerpetualTask(instanceSyncPerpetualTaskInfoDTO.getAccountIdentifier(),
           instanceSyncPerpetualTaskInfoDTO.getPerpetualTaskId());
       instanceSyncPerpetualTaskInfoDTO.setPerpetualTaskId(null);
-      instanceSyncPerpetualTaskInfoService.updatePerpetualTaskIdV1(instanceSyncPerpetualTaskInfoDTO);
+      instanceSyncPerpetualTaskInfoService.updatePerpetualTaskIdV1OrV2(instanceSyncPerpetualTaskInfoDTO);
     }
   }
